Revert NET-276 changes that break the binary API Will be re-applied in 3.0
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/net/trunk@1032805 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/net/nntp/ArticlePointer.java
@@ -28,7 +28,7 @@ package org.apache.commons.net.nntp;
 public final class ArticlePointer
 {
     /** The number of the referenced article. */
-    public long articleNumber;
+    public int articleNumber;
     /**
      * The unique id of the referenced article, including the enclosing
      * &lt and &gt symbols which are technically not part of the
